// products/src/apps/standalone/app/modules/products/src/src/shared/enums/product.enum.ts
export enum ProductEditorSections {
  Main = 'main',
  Content = 'content',
  Inventory = 'inventory',
  Category = 'category',
  Variants = 'variants',
  Attributes = 'attributes',
  Channels = 'channels',
  Shipping = 'shipping',
  Taxes = 'taxes',
  Recommendations = 'recommendations',
  // RecurringBilling = 'recurring_billing',
  Visibility = 'visibility',
}

export enum ProductTypes {
  Digital = 'digital',
  Physical = 'physical',
  Service = 'service',
}

export enum ChannelTypes {
  Pos = 'pos',
  Shop = 'shop',
  Market = 'market',
  Facebook = 'facebook',
  Mobilede = 'mobilede',
  Ebay = 'ebay',
  Autoscout24 = 'autoscout24',
  Amazon = 'amazon',
  Marketing = 'marketing',
  Instagram = 'instagram',
  FinanceExpress = 'finance_express',
  Link = 'link',
  Dropshipping = 'dropshipping',
}